Rename requests using a modal instead of inline
Everytime a request is renamed a new file is created
and the old file is removed. Before the request was
renamed using an inline edit which created some errors
and difficulties during validation. Now a modal is used
so that we can validate and rename only after the full name is changed.